requirements.txt
fastapi
uvicorn
ecdsa
pycryptodome
websockets
pydantic
pytest
pytest-asyncio
pytest-cov

